메인화면 완료
